Improvements to the Go, Rust, and Swift example codes after Magne's feedback.
- The Rust and Swift examples have been reworked to use interface
  extensions, and retroactive implementation of interfaces for
  intrinsic types, in order to showcase the full capabilities of these
  languages.

- Corresponding adjustments and improvements to the main text.

- Update to the acknowledgements section.

+++ Code/Go/coupled.go
@@ -4,7 +4,7 @@ import "fmt"
 
 func simple_sum(x []int32) int32 {
    var s int32
-   s = 0
+   s = int32(0)
    for i := 0; i < len(x); i++ {
      s += x[i]
    }
